Move check for file type (extension) to collection stage
File verification happens in two stages: collection of the files to be
verified, which is done in serial, and the verification itself, which is
(normally) done in parallel. In order to avoid wasting time on (and cluttering
the output with) irrelevant files, we check each file's name to see if the
extension, &c., indicate that it's an MP3. Previously, this was done in the
verification stage. This is an issue when a file is given directly on the
command line, because we might silently ignore what the user gave us. We
should always respect the user's wishes when they give us a file explicitly

Ideally all of the is-a-file and file-name checks should be done in parallel
for performance reasons. In the future maybe we can encode whether the file
type needs checked in a class or something and pass that to the verifier,
which would solve both problems. But this is probably fine for now

Fixes #1